Move javascript minification code to a new package
At a macro level, javascript minification didn't really make much sense
in Horde_Text_Filter to begin with.  It is not a text filter.  It was
put in there since at the time there was a single driver with no
options.  Now we have 4 different drivers, and some of these drivers
have advanced config options.

At a more micro level, this also allows us to do a couple of new, cool
things:

* For the backends that support (Closure, UglifyJS), we can now create
  sourcemaps automatically without any additional user action.
* Allows us to support user-created drivers directly from configuration
  without the need to edit any library code.
* Allows us to pass the entire list of JS files to parse to the driver,
  allowing each driver the opportunity to optimize parsing of these
  files (this is necessary for supporting sourcemaps).
* An added bonus -- although I don't have a specifc use case at this
  time -- is this allows for full user control over the javascript files
  that are served on any given page.  This could be accomplished by
  writing a local driver that extends any of the current drivers - now
  that a list of javascript files is provided, this driver could
  theoretically excise/add to that list.